+++ internal/service/platform/gitops/agent/data_source_gitops_agent.go
@@ -2,6 +2,7 @@ package agent
 
 import (
 	"context"
+	"fmt"
 
 	"github.com/antihax/optional"
 	hh "github.com/harness/harness-go-sdk/harness/helpers"
@@ -22,7 +23,9 @@ func DataSourceGitopsAgent() *schema.Resource {
 			"account_id": {
 				Description: "Account identifier of the GitOps agent.",
 				Type:        schema.TypeString,
-				Required:    true,
+				Computed:    true,
+				Optional:    true,
+				Deprecated:  "This field is deprecated and will be removed in a future release.",
 			},
 			"project_id": {
 				Description: "Project identifier of the GitOps agent.",
@@ -96,6 +99,21 @@ func DataSourceGitopsAgent() *schema.Resource {
 				Type:        schema.TypeString,
 				Computed:    true,
 			},
+			"with_credentials": {
+				Description: "Specify whether to retrieve the gitops agent's token. (The field agent_token will be populated only if the agent has never connected to Harness before). For retrieval of this information, the user associated to the token being used must have Gitops Agent Edit permissions",
+				Type:        schema.TypeBool,
+				Optional:    true,
+			},
+			"is_authenticated": {
+				Description: "This computed field specifies if the referenced agent ever successfully connected and was authenticated to harness. Note that this is different from whether the agent is currently connected. <b>Set with_credentials to true to allow computing of this field.</b> For retrieval of this information, the user associated to the token being used must have Gitops Agent Edit permissions",
+				Type:        schema.TypeBool,
+				Computed:    true,
+			},
+			"prefixed_identifier": {
+				Description: "Prefixed identifier of the GitOps agent. Agent identifier prefixed with scope of the agent",
+				Type:        schema.TypeString,
+				Computed:    true,
+			},
 		},
 	}
 	return resource
@@ -109,6 +127,7 @@ func dataSourceGitopsAgentRead(ctx context.Context, d *schema.ResourceData, meta
 	resp, httpResp, err := c.AgentApi.AgentServiceForServerGet(ctx, agentIdentifier, c.AccountId, &nextgen.AgentsApiAgentServiceForServerGetOpts{
 		OrgIdentifier:     optional.NewString(d.Get("org_id").(string)),
 		ProjectIdentifier: optional.NewString(d.Get("project_id").(string)),
+		WithCredentials:   optional.NewBool(d.Get("with_credentials").(bool)),
 	})
 
 	if err != nil && (httpResp == nil || httpResp.StatusCode != 404) {
@@ -123,5 +142,23 @@ func dataSourceGitopsAgentRead(ctx context.Context, d *schema.ResourceData, meta
 		return nil
 	}
 	readAgent(d, &resp)
+	err = readAgentDataSourceOnlyFields(d, &resp)
+	if err != nil {
+		return diag.FromErr(err)
+	}
+	return nil
+}
+
+func readAgentDataSourceOnlyFields(d *schema.ResourceData, agent *nextgen.V1Agent) error {
+	err := d.Set("with_credentials", d.Get("with_credentials"))
+	if err != nil {
+		return fmt.Errorf("error setting with_credentials field while reading response: %w", err)
+	}
+	if d.Get("with_credentials").(bool) {
+		err = d.Set("is_authenticated", agent.Credentials != nil && agent.Credentials.PrivateKey == "")
+		if err != nil {
+			return fmt.Errorf("error setting is_authenticated field while reading response: %w", err)
+		}
+	}
 	return nil
 }