+++ internal/service/platform/gitops/cluster/resource_gitops_cluster.go
@@ -28,8 +28,9 @@ func ResourceGitopsCluster() *schema.Resource {
 			"account_id": {
 				Description: "Account identifier of the GitOps cluster.",
 				Type:        schema.TypeString,
-				Required:    true,
-				ForceNew:    true,
+				Optional:    true,
+				Computed:    true,
+				Deprecated:  "This field is deprecated and will be removed in a future release.",
 			},
 			"project_id": {
 				Description: "Project identifier of the GitOps cluster.",